Made a lot of changes. Finally the backend is working now

+++ src/backend/prisma/migrations/20250919172700_expense_id_to_int/migration.sql
@@ -1,6 +1,6 @@
 -- CreateTable
 CREATE TABLE `Expense` (
-    `id` BIGINT NOT NULL AUTO_INCREMENT,
+    `id` INTEGER NOT NULL AUTO_INCREMENT,
     `amount` DECIMAL(10, 2) NOT NULL,
     `date` DATETIME(3) NOT NULL,
     `category` ENUM('Food', 'Groceries', 'Mobile_Bill', 'Travel', 'Shopping', 'Games', 'Subscription', 'EMI') NOT NULL,
